# Conflicts:
#	src/main/java/com/daylight/civilization_fire/common/content/item/agriculture/tool/ModHoeingToolItem.java

+++ src/main/java/com/daylight/civilization_fire/common/content/entity/bot/MiningBot.java
@@ -45,7 +45,7 @@ public final class MiningBot extends Bot {
     }
 
     @Override
-    public long getMaxEnergy() {
+    public int getMaxEnergy() {
         return 10000;
     }
